src/connection.py
from dotenv import load_dotenv
import os


load_dotenv()

class Config:
    HOST = os.getenv('HOST')
    USER = os.getenv('USER')
    PASSWORD = os.getenv('PASSWORD')
    DATABASE = os.getenv('DATABASE')