requirements.txt
Flask==3.1.0
mysql-connector-python==9.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
requests==2.25.1
python-dotenv==1.0.1